Extract ComparisonItem family to shared lib (Plan A remainder)
ComparisonItemBase + ProcedureStatsComparisonItem +
QueryStatsComparisonItem move to PerformanceMonitor.Ui; per-app copies
deleted. QueryStatsComparisonItem adopts Dashboard's superset (3 extra
nullable props, additive). No XAML changes (all bindings resolve by
property name). ActualPlanExecutor left per-app -- descoped: it
transitively depends on the per-app, drifted ReproScriptBuilder.
Completes the Plan A consolidation tail.
Plan: ~/.claude/plans/plan-a-remainder.md.

No deviations. (Consumer refs were Models.-qualified rather than relying
on a using; resolved by adding `using PerformanceMonitor.Ui;` and
dropping the Models. qualifier on the two moved types -- compiler-driven,
behaviour-preserving.)
