+++ PerformanceMonitor.Ui/ProcedureStatsComparisonItem.cs
@@ -1,12 +1,12 @@
 /*
  * Copyright (c) 2026 Erik Darling, Darling Data LLC
  *
- * This file is part of the SQL Server Performance Monitor Lite.
+ * This file is part of the SQL Server Performance Monitor.
  *
  * Licensed under the MIT License. See LICENSE file in the project root for full license information.
  */
 
-namespace PerformanceMonitorLite.Models;
+namespace PerformanceMonitor.Ui;
 
 public class ProcedureStatsComparisonItem : ComparisonItemBase
 {
